[change]: Data type int to ssize_t

+++ 0x15-file_io/2-append_text_to_file.c
@@ -11,7 +11,7 @@ int append_text_to_file(const char *filename, char *text_content)
 {
 	int fd;
 	int len;
-	int writep;
+	ssize_t writep;
 
 	if (filename == NULL)
 		return (-1);
